finetune script updated

+++ MU-LLaMA/engine_finetune.py
@@ -34,7 +34,7 @@ def train_one_epoch(model: LLaMA_adapter,
             lr_sched.adjust_learning_rate(optimizer, data_iter_step / len(data_loader) + epoch, args)
 
         imgs = imgs.to(device, non_blocking=True)
-        with torch.cuda.amp.autocast():
+        with torch.cuda.amp.autocast(dtype=torch.bfloat16):
              c_loss, m_loss = model(examples, labels, imgs)
         loss = c_loss  + m_loss * 0
         loss_value = loss.item()
@@ -67,7 +67,7 @@ def train_one_epoch(model: LLaMA_adapter,
             """
             epoch_1000x = int((data_iter_step / len(data_loader) + epoch) * 1000)
             log_writer.add_scalar('c_train_loss', c_loss_value_reduce, epoch_1000x)
-            log_writer.add_scalar('m_train_loss', m_loss_value_reduce, epoch_1000x)
+            log_writer.add_scalar('m_train_loss', m_loss_value_reduce.float(), epoch_1000x)
             log_writer.add_scalar('lr', lr, epoch_1000x)
